Fix memory leak
Parser combinators in Scala 2.10 leak. Quick history: in Scala 2.9 parsers had a state
and were not thread safe. In Scala 2.10 someone fixed this by moving this state as a
dynamically scoped variable which now cause this leak.

Because parsers are now thread safe, we can avoid creating a new parser each time.

Added a test case to ensure that the our lib is thread safe.

Change-Id: I2dfe43aeb8c78a61242d68bd65da018462be7e0d